On exception, return DONE instead of OK
We might need to revert this as I'm unsure of the side effects
this might have on functionality.  Returning DONE will cause
the server to close this thread without further processing.

Returning OK will proceed on to other modules, which in the case of
redirects, termination, and other errors seems like a bad idea.